Move fireap/string => fireap/util/string
