+++ lib/fireap/util/string.rb
